Refactor package structure and update OAuth2 user handling
