+++ build.gradle
@@ -37,16 +37,18 @@ dependencies {
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
-	compileOnly 'com.google.api-client:google-api-client:1.32.1'
 	implementation 'io.github.cdimascio:java-dotenv:5.2.2'
 	implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:2.6.0"
+	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+
+	implementation "io.jsonwebtoken:jjwt-api:0.11.5"
+	runtimeOnly "io.jsonwebtoken:jjwt-impl:0.11.5"
+	runtimeOnly "io.jsonwebtoken:jjwt-jackson:0.11.5"
 
 	testImplementation("org.junit.jupiter:junit-jupiter-api:5.11.4")
 	testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine:5.11.4")
 	testImplementation("org.mockito:mockito-core:5.12.0")
 	testImplementation("org.mockito:mockito-junit-jupiter:5.17.0")
-
-
 
 	// flyway
 	implementation 'org.flywaydb:flyway-mysql'
