--- src/main/java/com/verdict/verdict/config/PersistenceConfig.java
@@ -1,9 +1,0 @@
-package com.verdict.verdict.config;
-
-import org.springframework.beans.factory.annotation.Configurable;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configurable
-@EnableJpaAuditing
-public class PersistenceConfig {
-}
