--- src/main/java/com/verdict/verdict/dto/oauth2/GoogleOAuth2UserInfo.java
@@ -1,27 +1,0 @@
-package com.verdict.verdict.dto.oauth2;
-
-import java.util.Map;
-
-public class GoogleOAuth2UserInfo extends AbstractOAuth2UserInfo {
-
-    public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
-
-        super(attributes);
-    }
-    /**
-     * @return "google"
-     */
-    @Override
-    public String getProviderCode() {
-        return (String) attributes.get(ProviderInfo.GOOGLE.getProviderCode());
-    }
-
-    /**
-     * @return "user@gmail."
-     */
-    @Override
-    public String getUserIdentifier() {
-
-        return (String) attributes.get(ProviderInfo.GOOGLE.getIdentifier());
-    }
-}
