--- src/main/java/com/verdict/verdict/dto/response/PostExampleResponse.java
@@ -1,11 +1,0 @@
-package com.verdict.verdict.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class PostExampleResponse {
-
-    private String text;
-}
