+++ src/main/java/com/verdict/verdict/global/config/PersistenceConfig.java
@@ -1,0 +1,9 @@
+package com.verdict.verdict.global.config;
+
+import org.springframework.beans.factory.annotation.Configurable;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configurable
+@EnableJpaAuditing
+public class PersistenceConfig {
+}
