+++ src/main/java/com/verdict/verdict/global/info/OAuth2UserInfo.java
@@ -1,0 +1,19 @@
+package com.verdict.verdict.global.info;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.Map;
+
+@Getter
+@AllArgsConstructor
+public abstract class OAuth2UserInfo {
+
+
+    protected Map<String, Object> attributes;
+
+    public abstract String getProviderCode();
+
+    public abstract String getUserIdentifier();
+
+}
