+++ src/main/java/com/verdict/verdict/global/response/PostExampleResponse.java
@@ -1,0 +1,11 @@
+package com.verdict.verdict.global.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class PostExampleResponse {
+
+    private String text;
+}
