+++ src/main/java/com/verdict/verdict/global/signOut/SignoutRepository.java
@@ -1,0 +1,10 @@
+package com.verdict.verdict.global.signOut;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+
+
+public interface SignoutRepository extends JpaRepository<Signout, Long> {
+    @Query("select s from Signout s where s.identifier = :identifier")
+    Signout findByIdentifier(String identifier);
+}
