--- src/main/java/com/verdict/verdict/service/signOut/SignoutRepository.java
@@ -1,10 +1,0 @@
-package com.verdict.verdict.service.signOut;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-
-public interface SignoutRepository extends JpaRepository<Signout, Long> {
-    @Query("select s from Signout s where s.identifier = :identifier")
-    Signout findByIdentifier(String identifier);
-}
